Increase temporary buffer size of ibm,client-architecture-support call
The buffer size that we currently use for the client-architecture-
support call currently only works if there is not more than 1 TiB
of hot-pluggable RAM for the guest. This will likely not be enough
in the near future, so increase the buffer size from 128 kiB to 2 MiB
instead. That should be enough to accomodate for 16 TiB of hot-pluggable
memory.
While we're at it, also rename the "size" variable to something more
specific (since it is put into the namespace of the root node), and
add a proper error message in case the alloc-mem fails.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>