Improve F12 key handling in boot menu
When hitting F12 during boot multiple times (to make sure that you get
to the boot menu), it is sometimes impossible to select a menu entry
from the boot list - you always get the output "Invalid choice!" and
are dropped to the SLOF prompt.

The "boot-start" function already has some code at the beginning to
flush the keyboard input to get rid of multiple F12 keys. But if you
accidentially press F12 again after that code has been executed,
the code that checks the input for the selected boot menu entry
can not handle it and gets confused.

Fix this issue by flushing the keyboard input queue as late as possible
(i.e. after printing the boot menu items, since that could also take
a short while when the list is long and when using slow VGA text
output), and by ignoring all key sequences that start with an ESC
(like the F12 key ESC sequence) in the main keyboard input loop there.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>