usb-hid: Fix non-working comma key
Commit c8bd0cdce2bd ("Caps is not always shift") introduced
a global "ctrl" variable and two new cases to handle the shift
key in check_key_code() in usb-hid.c. However, one of these cases
uses the key code for the "," key and thus prevents that key
from working in the VGA console. The global ctrl variable that
is set by these two cases is never read again, so this is
apparently completely useless code, likely just an accidential
leftover from debugging the original issue. So let's simply
remove that bad code to get the comma key working again.

Fixes: c8bd0cdce2bd7c4498d61079553c89747656614d
Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>