Remove misleading padding fields from ROM header definition
The first two bytes of the date field are unused - and the date
is indeed encoded in the following six bytes, i.e. also using
the padding bytes. So the header definition is currently very
confusing. Let's simply remove the padding fields and make the
date field a little bit bigger instead - and add a proper comment
about the encoding of the date field.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Reviewed-by: Adrian Reber <areber@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>