Remember execution tokens of "write" and "read" for socket operations
The recv() and send() functions are currently assuming that the
dictionary of the NIC node is available, so that they can call
"read" or "write" directly. While this is working fine when only
doing network loading from the Open Firmware prompt, this does
not work anymore when a client like yaboot is trying to load
additional files.
Fix this issue by looking up the execution tokens of "read" and
"write" while opening the socket, so we can use these XT values
during recv() and send() later.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>