scsi: Remove debug functions from scsi-loader.fs
The ".context" function is just a colorful, but otherwise useless
and unused wrapper around "context @", so this can be removed
safely. Also the .ansi-* functions are way too fancy for just
colorizing the output of the debug function ".wordlist" in a
different color than the default one, so remove these .ansi-*
functions, too.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>