virtio-scsi: Fix descriptor order for SCSI WRITE commands
Virtio needs the descriptors ordered: All descriptors for buffers that
contain data that should go *to* the device have to be put into the
queue first. The descriptors for buffers that read *from* the device
have to be sorted in last. For SCSI WRITE commands (which we never
used in the past), the order was wrong, so QEMU complains with a
"Incorrect order for descriptors" message as soon as we use it for
SCSI WRITE commands.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>