bootmsg: Fix message for detected kernel
126e458fdc00 assumes that checkpoints are always printed from
the beginning of a line which is not the case when SLOF detects
a kernel image in check-boot-from-ram. When this happens, SLOF prints
a diagnostic message:
Detected RAM kernel at 400000 (180c3bc bytes)

Without 126e458fdc00, SLOF prints another checkpoint at the end of
the message, then it clears the last checkpoint with ."      " and
we see a nice message. With 126e458fdc00, we see:
     ted RAM kernel at 400000 (180c3bc bytes) C08FF

This adds "cr" to prevent the message from spoiling.

Fixes: 126e458fdc00 ("libbootmsg: Do not use '\b' characters when printing checkpoints")
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: Thomas Huth <thuth@redhat.com>